+++ grammar/fr/src/rules.rs
@@ -981,44 +981,31 @@ pub fn rules_time(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
                  time.value().form_time_of_day()?.is_12_clock()
              )
     );
-    b.rule_1_terminal("dd/-.mm/-.yyyy",
-        b.reg(r#"(3[01]|[12]\d|0?[1-9])[-/.](1[0-2]|0?[1-9])[-/.](\d{2,4})"#)?,
-        |text_match| helpers::year_month_day(
-            text_match.group(3).parse()?,
-            text_match.group(2).parse()?,
-            text_match.group(1).parse()?
-        )
+    // Written dates in numeric formats
+    b.rule_1_terminal("yyyy-mm-dd - ISO",
+                      b.reg(r#"(\d{4})[-/](0?[1-9]|1[0-2])[-/](3[01]|[12]\d|0?[1-9])"#)?,
+                      |text_match| helpers::year_month_day(
+                          text_match.group(1).parse()?,
+                          text_match.group(2).parse()?,
+                          text_match.group(3).parse()?)
     );
-    b.rule_1_terminal("yyyy-mm-dd",
-        b.reg(r#"(\d{2,4})-(1[0-2]|0?[1-9])-(3[01]|[12]\d|0?[1-9])"#)?,
-        |text_match| helpers::year_month_day(
-            text_match.group(1).parse()?,
-            text_match.group(2).parse()?,
-            text_match.group(3).parse()?
-        )
+    // Supporting these date formats also with whitespace as a separator for legacy
+    // But this seems too permissive?
+    b.rule_1_terminal("dd/mm/yy or dd/mm/yyyy",
+                      b.reg(r#"(0?[1-9]|[12]\d|3[01])[-\./ ](0?[1-9]|1[0-2])[-\./ ](\d{2,4})"#)?,
+                      |text_match| helpers::year_month_day(
+                          text_match.group(3).parse()?,
+                          text_match.group(2).parse()?,
+                          text_match.group(1).parse()?,
+                      )
     );
-    b.rule_1_terminal("dd/-mm",
-        b.reg(r#"(3[01]|[12]\d|0?[1-9])[/-](1[0-2]|0?[1-9])"#)?,
-        |text_match| helpers::month_day(
-            text_match.group(2).parse()?,
-            text_match.group(1).parse()?,
-        )
+    b.rule_1_terminal("dd/mm",
+                      b.reg(r#"(0?[1-9]|[12]\d|3[01])[\./ ](1[0-2]|0?[1-9])"#)?,
+                      |text_match| helpers::month_day(
+                          text_match.group(2).parse()?,
+                          text_match.group(1).parse()?)
     );
-    b.rule_1_terminal("dd mm yyyy",
-        b.reg(r#"(3[01]|[12]\d|0?[1-9]) (1[0-2]|0?[1-9]) (\d{2,4})"#)?,
-        |text_match| helpers::year_month_day(
-            text_match.group(3).parse()?,
-            text_match.group(2).parse()?,
-            text_match.group(1).parse()?
-        )
-    );
-    b.rule_1_terminal("dd mm",
-        b.reg(r#"(3[01]|[12]\d|0?[1-9]) (1[0-2]|0?[1-9])"#)?,
-        |text_match| helpers::month_day(
-            text_match.group(2).parse()?,
-            text_match.group(1).parse()?,
-        )
-    );
+    // End of Written dates in numeric formats
     b.rule_1_terminal("matin",
         b.reg(r#"mat(?:in[ée]?e?)?"#)?,
         |_| Ok(helpers::hour(4, false)?