+++ grammar/ja/src/training.rs
@@ -163,6 +163,8 @@ pub fn examples_time(v: &mut Vec<::rustling::train::Example<Dimension>>) {
     example!(v, check_moment!(c, [2013, 3, 8]), "女性の日");
     example!(v, check_moment!(c, [2013, 2, 15]), "三日後");
     example!(v, check_moment!(c, [2013, 2, 9]), "三日前");
+    example!(v, check_moment!(c, [2018, 1, 23]), "2018.1.23");
+    // TODO: date as "2018年1月23日"
 
     example!(v, check_moment_span!(c, [2013, 8, 1], [2013, 8, 11]), "八月の上旬");
     example!(v, check_moment_span!(c, [2013, 10, 1], [2013, 10, 11]), "十月の上旬");